code: Refactoring
